--- a/src/random/mersenne_twister.h
+++ b/src/random/mersenne_twister.h
@@ -42,12 +42,8 @@ class MersenneTwister : public RandomGenerator
   double sample() { return unif_(mt_); }
 
  protected:
-  //Not faster than using the quantile transformation in random_generator!
-  //double sampleUnitExponential() { return expo_(mt_); };
-
-  std::mt19937 mt_; 
+  std::mt19937_64 mt_; 
   std::uniform_real_distribution<> unif_;
-  std::exponential_distribution<> expo_;
 
  private:
   size_t generateRandomSeed() const;