Use the 64bit version instead of the 32bit one for the mersenne twister, as it should be faster on 64bit systems.