@paulstaab 2015-03-09
Makes a pretty small difference in runtimes (if any):

Command | Equal results | time master | time use_mt19937_64 | memory master | memory use_mt19937_64
------- | ------------- | ----------- | -------- | ------------- | ---------
scrm 100 100 -t 5 -T | No | 0.1s | 0.1s | 1658kb | 1653kb
scrm 20 100 -r 10 100 -t 3 -T -L | No | 0.075s | 0.074s | 1670kb | 1664kb
scrm 4 10000 -r 10 100 -t 5 | No | 0.53s | 0.51s | 1560kb | 1554kb
scrm 20 2 -r 500 100 -L | No | 0.6s | 0.61s | 3325kb | 3362kb
scrm 2000 1 -r 10 100 -l 50 -L | No | 0.52s | 0.52s | 1963kb | 1907kb
scrm 20 1 -r 4000 10000000 -l 300000 -T | No | 0.87s | 0.87s | 16449kb | 16467kb
scrm 10 100 -r 10 100 -I 3 3 3 4 0.5 -eN 0.1 0.05 -eN 0.2 0.5 | Yes | 0.045s | 0.044s | 1508kb | 1506kb
scrm 50 100 -r 20 200 -G 1.5 -l 100 | Yes | 0.041s | 0.04s | 1527kb | 1519kb

My tests show that using `boost/random` should be even faster. But given the tiny impact of this change, I guess that the expected speedup does not justify the additional decency.